Move addAliases() to separate file
